requirements.txt
numpy>=1.24.0
sympy>=1.12
customtkinter>=5.2.0
pytest>=7.0.0
matplotlib>=3.8.0
Pillow>=10.0.0